# Mark []byte passed to Sum64 as noescape