--- a/xxhash_amd64.go
+++ b/xxhash_amd64.go
@@ -4,6 +4,7 @@
 
 package xxhash
 
+//go:noescape
 func sum64(b []byte) uint64
 
 func writeBlocks(x *xxh, b []byte) []byte